opening scene cleanup
I still have some cleanup and touches to add, but I’m getting hungry
now and need a break. Here’s what this branch adds:

-added code to make sure the protest scene actually ENDS after a
certain amount of time

-the EnemySpawnerObject is set to inactive. it is activated in the
ProtestScene code once the protest has ended. Additionally, I doubled
the random range within which enemies can spawn

-FirstEnemy spawns after the scene fades in to view

-added RadioIntro sound to the canvas
