+++ Hold On Prototype/Assets/Audio/Radio Intro.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 298512b2e9a8b4b6fa37558e3614f9ff
+timeCreated: 1460836805
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
