+++ Hold On Prototype/Assets/Scripts/EnemySpawner.cs
@@ -8,20 +8,19 @@ public class EnemySpawner : MonoBehaviour {
 	private Vector3 spawnPoint;
 	public GameObject[] totalEnemies;
 
-	// Update is called once per frame
 	void Update () {
 		totalEnemies = GameObject.FindGameObjectsWithTag ("Enemy");
 		amount = totalEnemies.Length;
 
-		if (amount < 4) {
-			InvokeRepeating ("spawnEnemy", 5, 12f);
+		if (amount < 5) {
+			InvokeRepeating ("spawnEnemy", 5, 15f);
 		}
 	}
 
 	void spawnEnemy(){
-		spawnPoint.x = Random.Range(-20, 20);
+		spawnPoint.x = Random.Range(-40, 40);
 		spawnPoint.y = 1.5f;
-		spawnPoint.z = Random.Range(-10, 20);
+		spawnPoint.z = Random.Range(-20, 40);
 
 		Instantiate(enemy, spawnPoint, Quaternion.identity);
 		CancelInvoke();
