+++ Hold On Prototype/Assets/Scripts/GameOpening/FirstEnemyController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 71a7366569c2b40398288aa0989fb2fb
+timeCreated: 1460566885
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
